% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_shoreline.R
\name{crop_shoreline}
\alias{crop_shoreline}
\title{Crop shorelines}
\usage{
crop_shoreline(x, w = -180, e = 180, s = -90, n = 90)
}
\arguments{
\item{x}{shoreline data, output by \code{\link{read_shoreline}}.}

\item{w, e, s, n}{coordinates of the region to crop.}
}
\value{
The shorelines as a data.frame, with columns lon and lat
}
\description{
Crop shorelines, dealing with crossing of the date line, and return a data.frame
}
\examples{
# NB: read a very reduced version of the data, for tests
shorepath <- system.file("extdata", ".", package="shoreliner")
w <- read_shoreline("c", 1, path=shorepath)
plot(w, max.plot=1)

head(crop_shoreline(w, -50, 50, -50, 50))

plot(crop_shoreline(w), type="l")
plot(crop_shoreline(w, -50, 50, -50, 50), type="l")
plot(crop_shoreline(w, -170, 170, -50, 50), type="l")
plot(crop_shoreline(w, 140, 290, -50, 0), type="l")
plot(crop_shoreline(w, -200, -140, 40, 90), type="l")
\dontrun{
ggplot(crop_shoreline(w, 140, 290, -50, 0)) +
  geom_polygon(aes(lon, lat)) + coord_map()
}
}
