% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shoreline.R
\name{download_shoreline}
\alias{download_shoreline}
\title{Download GSHHG data}
\usage{
download_shoreline(path = ".", set.options = NULL, server = "noaa")
}
\arguments{
\item{path}{path to a directory where the data will be downloaded. By default the current directory.}

\item{set.options}{logical, when TRUE the path to the data is set as an option in ~/.Rprofile, so that new R sessions can find it. By default, the user is prompted in interactive sessions or it is set to `FALSE` in non-interactive ones.}

\item{server}{string identifying the server to download the data from; either `noaa` or `hawaii`.}
}
\value{
The path to the uncompressed data, invisibly.
}
\description{
Download GSHHG data as shapefiles and uncrompress them.
}
\examples{
\dontrun{
download_shoreline()}
}
