% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shoreline.R
\name{read_shoreline}
\alias{read_shoreline}
\title{Read shoreline from GSHHG}
\usage{
read_shoreline(resolution, level, path = getOption("shoreline.path"))
}
\arguments{
\item{resolution}{resolution at which the data should be read; either the full name, an abbreviation, or a number.
\itemize{
\item 1 : cruise (c)
\item 2 : low (l)
\item 3 : intermediate (i)
\item 4 : high (h)
\item 5 : full (f)
}}

\item{level}{level \strong{down to} which the data should be read
\itemize{
\item 1 : boundary between land and ocean
\item 2 : boundary between lake and land
\item 3 : boundary between island-in-lake and lake
\item 4 : boundary between pond-in-island and island
}}

\item{path}{path to the GSHHG data. By defautl this is set by an option. If the data is not found, it will be downloaded.}
}
\value{
A "simple feature" (\code{\link[sf]{sf}}) object.
}
\description{
Read shoreline from GSHHG
}
\examples{
shorepath <- system.file("extdata", ".", package="shoreliner")
# NB: this is a very reduced version of the data, only for tests

# various abbreviations can be used
w <- read_shoreline("cruise", 1, path=shorepath)
w <- read_shoreline("c", 1, path=shorepath)
w <- read_shoreline(1, 1, path=shorepath)
plot(w, max.plot=1)

# the resulting objet can be manipulated with all the functions in sf
sf::st_bbox(w)
plot(sf::st_crop(w, xmin=-10, xmax=50, ymin=30, ymax=70), max.plot=1)
# etc.
}
